Make SRTPCipherCTR an abstract class and SRTPCipherCTRJava the first implementation
factor process() arguments checks into checkProcessArgs
